- Abstracting Table component into its own class for reuse
