+++ packages/satisfactory/src/components/TableItems.tsx
@@ -1,20 +1,18 @@
-﻿import { ReactNode, useEffect, useState } from "react";
+﻿import { useEffect, useState } from "react";
 import {
 	Cell,
 	createColumnHelper,
-	flexRender,
-	getCoreRowModel, getFilteredRowModel, getSortedRowModel, Header, SortingState,
+	getCoreRowModel, getFilteredRowModel, getSortedRowModel, SortingState,
 	useReactTable,
 } from "@tanstack/react-table";
 import { loadInventory } from "@/api/inventory.ts";
 import { CellItemName } from "@/components/CellItemName.tsx";
 import { FieldText } from "@/components/FieldText.tsx";
-import { IconSort } from "@/components/Icons.tsx";
+import { TableData } from "@/components/TableData.tsx";
 import { ViewItem } from "@/components/ViewItem.tsx";
 import { IInventoryItem } from "@/types.ts";
 
 const columnHelper = createColumnHelper<IInventoryItem>();
-const DefaultCellCls = "border-r border-b first:border-l px-2 py-1";
 const DefaultColumns = [columnHelper.accessor("name", {
 	header: "Name",
 	cell: (info) => <CellItemName cell={info.cell} />,
@@ -42,7 +40,6 @@ const DefaultColumns = [columnHelper.accessor("name", {
 })];
 
 export function TableItems() {
-	const [columns] = useState(DefaultColumns);
 	const [data, setData] = useState<IInventoryItem[]>([]);
 	const [search, setSearch] = useState("");
 	const [globalFilter, setGlobalFilter] = useState<string>();
@@ -53,8 +50,8 @@ export function TableItems() {
 		desc: false,
 	}]);
 	const table = useReactTable({
-		columns,
 		data,
+		columns: DefaultColumns,
 		globalFilterFn: "includesString",
 		getCoreRowModel: getCoreRowModel(),
 		getFilteredRowModel: getFilteredRowModel(),
@@ -66,72 +63,9 @@ export function TableItems() {
 			globalFilter,
 		},
 	});
-	const tableHeaders = table.getHeaderGroups().map((headerGroup) => {
-		const rows = headerGroup.headers.map((header) => {
-			const { column } = header;
-			let title = "";
-			let sortIcon: ReactNode;
-			if (column.getCanSort()) {
-				const sortDir = column.getIsSorted();
-				switch (column.getNextSortingOrder()) {
-					case "asc":
-						title = "Sort Ascending";
-						break;
-					case "desc":
-						title = "Sort Descending";
-						break;
-					default:
-						title = "Clear Sort";
-				}
-				if (sortDir) {
-					const sortCls = ["size-6"];
-					if (sortDir === "asc") {
-						sortCls.push("rotate-180 -scale-x-90");
-					}
-					sortIcon = (
-						<div className="absolute right-2 top-2">
-							<IconSort className={sortCls.join(" ")} />
-						</div>
-					);
-				}
-			}
-			return (
-				<th
-					key={header.id}
-					className={getHeaderClass(header)}
-					onClick={column.getToggleSortingHandler()}
-					title={title}
-				>
-					<span>
-						{flexRender(header.column.columnDef.header, header.getContext())}
-					</span>
-					{sortIcon}
-				</th>
-			);
-		});
-		return (
-			<tr key={headerGroup.id}>
-				{rows}
-			</tr>
-		);
-	});
 
 	function onChangeSearch(searchValue: string) {
 		table.setGlobalFilter(searchValue);
-	}
-
-	function getHeaderClass(header: Header<IInventoryItem, unknown>) {
-		const cls = ["z-1 min-w-64 p-2 border-r first:border-l border-y sticky top-0"];
-		if (header.column.getCanSort()) {
-			cls.push("cursor-pointer select-none hover:bg-blue-300");
-		}
-		if (header.column.getIsSorted()) {
-			cls.push("bg-blue-400");
-		}
-		else {
-			cls.push("bg-gray-300");
-		}
-		return cls.join(" ");
 	}
 
 	function onClickCell(cell: Cell<IInventoryItem, unknown>) {
@@ -139,36 +73,6 @@ export function TableItems() {
 			setActiveCell(cell.row.original);
 			setShowItemDialog(true);
 		}
-	}
-
-	function getCellClass(cell: Cell<IInventoryItem, unknown>) {
-		const cls = [DefaultCellCls];
-		const { column } = cell;
-		const { meta } = column.columnDef;
-		if (column.id === "total") {
-			const value = cell.getValue<number>();
-			if (value < 0) {
-				cls.push("bg-red-200");
-			}
-			else if (value > 0) {
-				cls.push("bg-green-200");
-			}
-			else {
-				cls.push("bg-white");
-			}
-		}
-		else {
-			cls.push("bg-white");
-		}
-		if (meta) {
-			if (meta.cellCls) {
-				cls.push(meta.cellCls);
-			}
-			if (meta.canClick) {
-				cls.push("cursor-pointer hover:bg-blue-200");
-			}
-		}
-		return cls.join(" ");
 	}
 
 	useEffect(() => {
@@ -186,28 +90,10 @@ export function TableItems() {
 					onInputChange={onChangeSearch}
 				/>
 			</section>
-			<section className="overflow-auto flex-1">
-				<table className="border-spacing-0 border-separate w-full">
-					<thead>
-						{tableHeaders}
-					</thead>
-					<tbody>
-						{table.getRowModel().rows.map((row) => (
-							<tr key={row.id}>
-								{row.getVisibleCells().map((cell) => (
-									<td
-										key={cell.id}
-										className={getCellClass(cell)}
-										onClick={() => onClickCell(cell)}
-									>
-										{flexRender(cell.column.columnDef.cell, cell.getContext())}
-									</td>
-								))}
-							</tr>
-						))}
-					</tbody>
-				</table>
-			</section>
+			<TableData<IInventoryItem>
+				table={table}
+				onClickCell={onClickCell}
+			/>
 			<ViewItem
 				record={activeCell}
 				show={showItemDialog}
